requirements.txt
Django==2.0.2 # BSD2
django-bootstrap4==0.0.5 # BSD3
etcd3==0.7.0 # Apache-2.0
grpcio==1.9.1 # BSD3
monotonic==1.4 # Apache-2.0
numpy==1.14.0 # BSD
protobuf==3.5.1 # BSD
pytz==2017.3 # MIT
six==1.11.0  # MIT
tenacity==4.8.0 # Apache-2.0
django-angular==2.0.2 # MIT
django-cors-headers==2.2.0 # MIT
djangorestframework==3.7.7 # MIT
drf-nested-routers==0.90.2 # Apache-2.0
